finish proyect

+++ src/main/java/com/record/company/com/bussines/servicesImpl/AlbumFileServicesImpl.java
@@ -2,20 +2,34 @@ package com.record.company.com.bussines.servicesImpl;
 
 import com.record.company.com.bussines.IAlbumFileServices;
 import com.record.company.com.domain.dto.albumFile.AlbumFileDto;
+import com.record.company.com.domain.dto.albumFile.AlbumFileUrlDto;
 import com.record.company.com.domain.entity.AlbumFile;
 import com.record.company.com.exception.NotFoundException;
 import com.record.company.com.persistence.repository.IAlbumFileRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.UrlResource;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.core.io.Resource;
 
 import java.io.File;
+
+import java.io.IOException;
+import java.io.NotActiveException;
+import java.net.MalformedURLException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Service
 public class AlbumFileServicesImpl implements IAlbumFileServices {
@@ -26,15 +40,32 @@ public class AlbumFileServicesImpl implements IAlbumFileServices {
     @Autowired
     ModelMapper modelMapper;
 
-    private final String FILE_PATH = "E:\\Proyectos-Java\\record.company\\src\\main\\resources\\static\\image";
+    private Path rootLocation;
+
+    private final String FILE_PATH = "C:\\Users\\leona\\Desktop\\imagenes";
+
     @Override
     public List<AlbumFileDto> getAllAlbumFile() {
-        return null;
+        return albumFileRepository.findAll().stream().map(albumFile -> modelMapper.map(albumFile, AlbumFileDto.class)).collect(Collectors.toList());
     }
 
     @Override
-    public AlbumFileDto getAlbumFileById(int id) {
-        return null;
+    public AlbumFileDto getFileById(int id) {
+        return modelMapper.map(albumFileRepository.findById(id), AlbumFileDto.class);
+    }
+
+    @Override
+    public Resource getAlbumFileById(int id) throws MalformedURLException {
+        AlbumFile albumFile = albumFileRepository.findById(id).orElseThrow(() -> new NotFoundException("Id del album no encotrado"));
+
+        Path imagePath = Paths.get(albumFile.getPath(), albumFile.getNameImg());
+        Resource resource = new UrlResource(imagePath.toUri());
+
+        if (resource.exists() && resource.isReadable()) {
+            return resource;
+        } else {
+            throw new NotFoundException("La imagen no se encuentra");
+        }
     }
 
     @Override
@@ -51,48 +82,107 @@ public class AlbumFileServicesImpl implements IAlbumFileServices {
             long fileSize = file.getSize();
             long maxFileSize = 5 * 1024 * 1024;
 
-            if(fileSize > maxFileSize){
+            if (fileSize > maxFileSize) {
                 return null;
             }
 
-            if(!fileOriginalName.endsWith(".jpg") &&
+            if (!fileOriginalName.endsWith(".jpg") &&
                     !fileOriginalName.endsWith(".jpeg")
-            && !fileOriginalName.endsWith(".png")
+                    && !fileOriginalName.endsWith(".png")
 
-            ){
-                return null;
+            ) {
+                throw new NotFoundException("Imagen no compatible");
             }
 
             String fileExtension = fileOriginalName.substring(fileOriginalName.lastIndexOf("."));
             String newFileName = fileName + fileExtension;
 
+
             File folder = new File(FILE_PATH);
-            if(!folder.exists()){
+            if (!folder.exists()) {
                 folder.mkdirs();
             }
 
-            Path path = Paths.get(FILE_PATH +"/" + newFileName);
-            Files.write(path,bytes);
+            Path path = Paths.get(FILE_PATH + "/" + newFileName);
+            Files.write(path, bytes);
 
             albumFile.setPath(FILE_PATH);
             albumFile.setNameImg(newFileName);
             albumFile.setType(fileExtension);
+            albumFile.setUrlImageLocation("/api/v1/albumFile/" + newFileName);
 
-            return modelMapper.map(albumFileRepository.save(albumFile),AlbumFileDto.class);
-        }
-        catch (Exception e){
-           throw  new RuntimeException("Error save image");
+            return modelMapper.map(albumFileRepository.save(albumFile), AlbumFileDto.class);
+        } catch (Exception e) {
+            throw new RuntimeException("Error save image");
         }
     }
 
     @Override
     public AlbumFileDto updateAlbum(int id, MultipartFile file) {
-        return null;
+
+        Optional<AlbumFile> albumFileExist = albumFileRepository.findById(id);
+        AlbumFile albumFile = new AlbumFile();
+
+        if (albumFileExist.isPresent()) {
+            albumFile = albumFileExist.get();
+
+            try {
+                String fileName = UUID.randomUUID().toString();
+                byte[] bytes = file.getBytes();
+                String fileOriginalName = file.getOriginalFilename();
+
+
+                long fileSize = file.getSize();
+                long maxFileSize = 5 * 1024 * 1024;
+
+                if (fileSize > maxFileSize) {
+                    return null;
+                }
+
+                if (!fileOriginalName.endsWith(".jpg") &&
+                        !fileOriginalName.endsWith(".jpeg")
+                        && !fileOriginalName.endsWith(".png")
+
+                ) {
+                    throw new NotFoundException("Imagen no compatible");
+                }
+
+                String fileExtension = fileOriginalName.substring(fileOriginalName.lastIndexOf("."));
+                String newFileName = fileName + fileExtension;
+
+
+                File folder = new File(FILE_PATH);
+                if (!folder.exists()) {
+                    folder.mkdirs();
+                }
+
+                Path path = Paths.get(FILE_PATH + "/" + newFileName);
+                Files.write(path, bytes);
+
+                albumFile.setPath(FILE_PATH);
+                albumFile.setNameImg(newFileName);
+                albumFile.setType(fileExtension);
+                albumFile.setUrlImageLocation("/api/v1/albumFile/" + newFileName);
+
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+            return modelMapper.map(albumFileRepository.save(albumFile), AlbumFileDto.class);
+        }
+        throw new NotFoundException("Error al almacenar la imagen");
     }
 
     @Override
     public AlbumFileDto deleteAlbum(int id) {
-        
-        return null;
+        Optional<AlbumFile> albumFileExist = albumFileRepository.findById(id);
+        AlbumFile albumFile = new AlbumFile();
+
+        if(albumFileExist.isPresent()){
+            albumFile = albumFileExist.get();
+            albumFileRepository.deleteById(id);
+        }
+
+        return modelMapper.map(albumFile,AlbumFileDto.class);
     }
+
 }
